# api_documentation.yaml
openapi: 3.0.0
...
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/weAreWords/H.A.R.A.M.B.E/1.0.0
info:
  description: This api is used to track activities and users that interact with said activities
  version: "1.0.0"
  title: Activity Tracker API
  contact:
    email: huietj1@nku.eud
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /users:
    get:
      tags:
        - users
      summary: returns all Users
      operationId: getUsers
      responses:
        '200':
          description: Users found in database
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
        - users
      summary: Create a User with form data
      operationId: addUser
      requestBody:
         $ref: '#/components/requestBodies/UserCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 12
                  firstName:
                    type: string
                    example: "Jacob"
                  lastName:
                    type: string
                    example: "Huiet"
                  password:
                    type: string
                    example: "hashed password"
                  
                  status:
                    type: string
                    enum:
                      - admin
                      - user
                      - inactive
                    default: inactive
                    example: "inactive"
                  email:
                    type: string
                    example: ""
                  phone:
                    type: string
                    example: ""
                  createdAt:
                    type: string
                    example: '2019-05-31T17:21:55.302Z'
                  updatedAt:
                    type: string
                    example: '2019-05-31T17:21:55.302Z'
          description: User created
        '400':
          description: 'Err: [custom error message]'
...
  /users/{userId}:
    get:
      tags:
        - users
      summary: Gets a user by ID.
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid Id
        '404':
          description: User not found
    put:
      tags:
        - users
      summary: update a user based on id using form data
      operationId: updateUserById
      requestBody:
        $ref: '#/components/requestBodies/UserUpdate'
      parameters:
        - name: userId
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: '[custom error message]'
    delete:
      tags:
        - users
      summary: Deletes a user
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: User id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: user deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: User Id not found in database
...
  /activities/{activityId}/Users:
    parameters:
      - name: activityId
        in: path
        description: Activity ID
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - user activity
      summary: Get users involved with an activity
      operationId: getUsersAtActivity
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity_Attendance'
        '404':
          description: Invalid Activity Id
  /activites/{activityId}/User/{userId}:
    parameters:
      - name: activityId
        in: path
        description: Activity ID
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        description: User ID
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - user activity
      summary: Get info about a user attending an Activity
      operationId: getUserAtActivity
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity_Attendance'
        '404':
          description: Invalid Activity or User Id
    post:
      tags:
        - user activity
      summary: Set user status for an activity
      operationId: postUserAtActivity
      requestBody:
        $ref: '#/components/requestBodies/ActivityAttendancePost'
      responses:
        '201':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity_Attendance'
        '400':
          description: Invalid rsvp value entered. please enter yes, no, or maybe 
        '404':
          description: Invalid Activity or User Id
    put:
      tags:
        - user activity
      summary: Update user status and attendance for an activity
      operationId: putUserAtActivity
      requestBody:
        $ref: '#/components/requestBodies/ActivityAttendancePut'
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity_Attendance'
        '400':
          description: Invalid values entered for rsvp or attendance
        '404':
          description: Invalid Activity or User Id
    delete:
      tags:
        - user activity
      summary: Delete user status and attendance for an activity
      operationId: deleteUserAtActivity
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity_Attendance'
        '404':
          description: Invalid Activity or User Id
...
components:
  requestBodies:
    UserCreate:
      content:
        application/json:
          schema:
            type: object
            properties:
              firstName:
                type: string
                example: "Jacob"
              lastName:
                type: string
                example: "Huiet"
              password:
                type: string
                example: "hashed password"
              email:
                type: string
                example: "huietj1@gmail.com"
    UserUpdate:
      content:
        application/json:
          schema:
            type: object
            properties:
              firstName:
                type: string
                example: "Jacob"
              lastName:
                type: string
                example: "Huiet"
              password:
                type: string
                example: "hashed password"
              email:
                type: string
                example: "huietj1@gmail.com"
              status:
                type: string
                enum:
                  - admin
                  - user
                  - inactive
              phone:
                type: string
                example: "859-444-4444"
...
    ActivityAttendancePost:
      content:
        application/json:
          schema:
            type: object
            required:
              - rsvp
            properties:
              rsvp:
                type: string
                enum:
                  - yes
                  - no
                  - maybe
    ActivityAttendancePut:
      content:
        application/json:
          schema:
            type: object
            properties:
              rsvp:
                type: string
                enum:
                  - yes
                  - no
                  - maybe
              attendance:
                type: boolean
...
  schemas:
    User:
      type: object
      required:
        - firstName
        - lastName
        - password
        - email
      properties:
        id:
          type: integer
          example: 12
        firstName:
          type: string
          example: "Jacob"
        lastName:
          type: string
          example: "Huiet"
        password:
          type: string
          example: "hashed password"
        email:
          type: string
          example: "huietj1@gmail.com"
        status:
          type: string
          enum:
            - admin
            - user
            - inactive
          default: inactive
        phone:
          type: string
          example: "859-444-4444"
        createdAt:
          type: string
          example: '2019-05-31T17:21:55.302Z'
        updatedAt:
          type: string
          example: '2019-06-31T17:22:55.302Z'
...
    Activity_Attendance:
      type: object
      required:
        - userId
        - activityId
        - rsvp
      properties:
        userId:
          type: integer
        activityId:
          type: integer
        rsvp:
          type: string
          enum:
            - yes
            - no
            - maybe
        attendance:
          type: boolean
          default: false
...